[feat/WEEK-Stat] heatmapData 데이터 처리 위치 변경
